% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_texpackages.R
\name{get_texpackages}
\alias{get_texpackages}
\title{Get list of TeX packages installed on System}
\usage{
get_texpackages()
}
\value{
character
}
\description{
Fetch all TeX packages currently installed on system
}
\details{
If OS is Windows function checks against MikTex else function checks against TexLive.
}
\examples{
\dontrun{head(get_texpackages())}
}
\concept{utils}
