% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texPreview-deprecated.R
\name{getTexPackages}
\alias{getTexPackages}
\title{Get list of TeX packages installed on System \lifecycle{deprecated}}
\usage{
getTexPackages()
}
\description{
This function has been deprecated, use \link[=get_texpackages]{get_texpackages} instead.
}
