% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texPreview-deprecated.R
\name{buildUsepackage}
\alias{buildUsepackage}
\title{Build usepackage command for TeX document \lifecycle{deprecated}}
\usage{
buildUsepackage(pkg, options = NULL, uselibrary = NULL, chk.inst = FALSE)
}
\arguments{
\item{pkg}{character, name of TeX package}

\item{options}{character, name(s) of options to use in the package}

\item{uselibrary}{character, part of document preamble to specify a uselibrary call related to package}

\item{chk.inst}{logical, invokes a check to see if pkg is currently installed on system (default FALSE)}
}
\description{
This function has been deprecated, use \link[texPreview:build_usepackage]{build_usepackage} instead.
}
