% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildUsepackage.R
\name{buildUsepackage}
\alias{buildUsepackage}
\title{Build usepackage command for TeX document}
\usage{
buildUsepackage(pkg, opts = NULL, chk.inst = FALSE)
}
\arguments{
\item{pkg}{character, name of TeX package}

\item{opts}{character, name(s) of functions to call in the package}

\item{chk.pkg}{logical, invokes a check to see if pkg is currently installed on system}
}
\value{
character
}
\description{
input TeX package name and optional package functions to create usepackage call
}
\details{
opts is an optional argument, if it is NULL (default) then only the call for the package is returned.
See the TeX wikibook for more information \url{https://en.wikibooks.org/wiki/LaTeX/Document_Structure#Packages} on 
the usepackage command. If chk.pkg finds that the package is not installed on system function returns NULL.
}
\examples{
buildUsepackage(pkg = 'xcolor')
buildUsepackage(pkg = 'xcolor',opts = 'usenames')

#build many at once using mapply
mapply(buildUsepackage,pkg=list('xcolor','multrow','helvet','amsmath','rotating','graphicx'),
opts=list(c('usenames','dvispnames','svgnames','table'),NULL,NULL,NULL,NULL,NULL))
}

