% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTexPackages.R
\name{getTexPackages}
\alias{getTexPackages}
\title{Get list of TeX packages installed on System}
\usage{
getTexPackages()
}
\value{
data.frame
}
\description{
Fetch all TeX packages currently installed on system
}
\details{
If OS is Windows function checks against MikTex else function checks against TexLive.
}
\examples{
head(getTexPackages())
}

