% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texPreview.R
\name{texPreview}
\alias{texPreview}
\title{Render and Preview snippets of TeX in R Viewer}
\usage{
texPreview(obj, fileNM, fileDir = NULL, overwrite = T, imgFormat = "png",
  print.xtable.opts = list())
}
\arguments{
\item{obj}{character, TeX script}

\item{fileNM}{character, name to use in output files}

\item{fileDir}{character, output destination. If NULL a temp.dir() will be used and no output will be saved}

\item{overwrite}{logical, controls if overwriting of output fileNM* files given their existences}

\item{imgFormat}{character, defines the type of image the PDF is converted to.}

\item{print.xtable.opts}{list containing arguments to pass to print.table, relevant only if xtable is used as the input}
}
\description{
input TeX script into the function and it renders a pdf and converts it an image which is sent to Viewer.
}
\details{
The function assumes the system has pdflatex installed and it is defined in the PATH. The function does not return anything to R.
If fileDir is specified then two files are written to the directory. An image file of the name fileNM with the extension specified in imgFormat.
The default extension is png.The second file is the TeX script used to create the output of the name fileNM.tex.
}
\examples{
data('iris')

texPreview(obj = xtable(head(iris,10)),fileNM = 'eq',imgFormat = 'svg')

tex='\\\\begin{tabular}{llr}
\\\\hline
\\\\multicolumn{2}{c}{Item} \\\\\\\\
\\\\cline{1-2}
Animal    & Description & Price (\\\\$) \\\\\\\\
\\\\hline
Gnat      & per gram    & 13.65      \\\\\\\\
& each        & 0.01       \\\\\\\\
Gnu       & stuffed     & 92.50      \\\\\\\\
Emu       & stuffed     & 33.33      \\\\\\\\
Armadillo & frozen      & 8.99       \\\\\\\\
\\\\hline
\\\\end{tabular}'

texPreview(obj = x,fileNM = 'eq',imgFormat = 'png')
}

